// frontend/javascript/react_native/mobile_apps/mobile_app-solar-facil/src/types/BeneficiadoType.ts
// /src/types/BeneficiadoType.ts
export type BeneficiadoType = {
  nome: string;
  email: string;
  telefone: string;
  cpf: string;
  documentoIdentidade: string;
  dataNascimento: string;
  endereco: string;
  cidade: string;
  estado: string;
  nomeConcessionaria: string;
  consumoMedio: string;
  planoDesejado: string;
  aceitaTermos: boolean;
  observacoes: string;
};